+++ R/is_r_package.R
@@ -2,11 +2,13 @@
 #'
 #' Check whether a GitHub repository is for an R package or not.
 #' @param pkg Package name.
-#' @inheritParams echogithub::description_find
+#' @param owner Name of the owner of the R packages'
+#' GitHub repository (optional).
+#' @inheritParams echogithub::r_repos_data
 #' @returns boolean
 #'
 #' @importFrom utils installed.packages
-#' @importFrom echogithub r_repos_data description_find
+#' @importFrom echogithub r_repos_data description_extract
 #' @export
 #' @examples
 #' is_pkg <- is_r_package(pkg="stats")
@@ -20,14 +22,16 @@ is_r_package <- function(pkg,
     } else {
         pkgs <- echogithub::r_repos_data(add_downloads = FALSE,
                                          verbose = verbose)
-        if(pkg %in% pkgs$pkgs$r_repo){
+        if(pkg %in% unique(pkgs$package)){
             return(TRUE)
         } else if (is.null(owner)){
             stopper("Please provide owner argument to infer R package status.")
         } else {
-            d <- echogithub::description_find(owner = owner,
-                                              repo = pkg,
-                                              verbose = verbose)
+            d <- echogithub::description_extract(
+                ref = paste(owner,pkg,sep="/"),
+                fields = "Package",
+                verbose = verbose
+            )
             if(!is.null(d)) return(TRUE)
         }
     }